requirements.txt
asgiref==3.2.7
Babel==2.8.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-allauth==0.42.0
django-appconf==1.0.4
django-bootstrap4==1.1.1
django-cors-headers==3.3.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.2.0
django-filters==0.2.1
django-import-export==2.1.0
django-js-asset==1.2.2
django-measurement==3.2.3
django-mptt==0.11.0
django-phonenumber-field==4.0.0
django-rest-framework==0.1.0
django-select2==7.4.2
django-sql-explorer==1.1.3
django-tables2==2.3.1
django-tempus-dominus==5.1.2.13
django-versatileimagefield==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docopt==0.6.2
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
measurement==3.2.0
mpmath==1.1.0
num2words==0.5.10
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
psycopg2==2.8.5
PyPDF2==1.26.0
python-dateutil==2.8.1
python-magic==0.4.18
python-magic-bin==0.4.14
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
sympy==1.6
tablib==2.0.0
unicodecsv==0.14.1
urllib3==1.25.9
webencodings==0.5.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
